+++ app/state/applets/applets.thunks.js
@@ -1,4 +1,5 @@
 import { Actions } from 'react-native-router-flux';
+import * as firebase from 'react-native-firebase';
 import * as R from 'ramda';
 import {
   getApplets,
@@ -12,12 +13,12 @@ import {
   postAppletBadge,
   getTargetApplet,
   getAppletSchedule,
-} from '../../services/network';
-import { scheduleNotifications } from '../../services/pushNotifications';
+} from "../../services/network";
+import { scheduleNotifications } from "../../services/pushNotifications";
 // eslint-disable-next-line
 import { downloadResponses, downloadAppletResponses } from '../responses/responses.thunks';
 import { downloadAppletsMedia, downloadAppletMedia } from '../media/media.thunks';
-import { activitiesSelector } from './applets.selectors';
+import { activitiesSelector, allAppletsSelector } from './applets.selectors';
 import {
   replaceTargetAppletSchedule,
   setNotifications,
@@ -27,12 +28,16 @@ import {
   saveAppletResponseData,
   replaceTargetApplet,
   setDownloadingTargetApplet,
-} from './applets.actions';
-import { authSelector, userInfoSelector, loggedInSelector } from '../user/user.selectors';
-import { setCurrentApplet } from '../app/app.actions';
+} from "./applets.actions";
+import {
+  authSelector,
+  userInfoSelector,
+  loggedInSelector,
+} from "../user/user.selectors";
+import { setCurrentApplet } from "../app/app.actions";
 
-import { sync } from '../app/app.thunks';
-import { transformApplet } from '../../models/json-ld';
+import { sync } from "../app/app.thunks";
+import { transformApplet } from "../../models/json-ld";
 
 /* deprecated */
 export const scheduleAndSetNotifications = () => (dispatch, getState) => {
@@ -55,7 +60,7 @@ export const getInvitations = () => (dispatch, getState) => {
     });
 };
 
-export const getSchedules = appletId => (dispatch, getState) => {
+export const getSchedules = (appletId) => (dispatch, getState) => {
   const state = getState();
   const auth = authSelector(state);
 
@@ -69,6 +74,89 @@ export const getSchedules = appletId => (dispatch, getState) => {
     });
 };
 
+export const setReminder = () => async (dispatch, getState) => {
+  const state = getState();
+  const applets = allAppletsSelector(state);
+  const notifications = [];
+  
+  closeExistingNotifications();
+  applets.forEach(applet => {
+    const validEvents = applet.schedule.events.filter(event => event.valid);
+
+    validEvents.forEach(event => {
+      event.data.notifications.forEach(notification => {
+        const values = notification.start.split(':');
+        const date = new Date();
+
+        date.setHours(values[0]);
+        date.setMinutes(values[1]);
+        if (date.getTime() > Date.now()) {
+          notifications.push({
+            eventId: event.id,
+            appletId: applet.id.split('/').pop(),
+            activityId: event.data.activity_id,
+            activityName: event.data.title,
+            date: date.getTime()
+          });
+        }
+      })
+    })
+  });
+  
+  notifications.forEach(notification => {
+    const settings = { showInForeground: true };
+    const AndroidChannelId = 'MindLoggerChannelId';
+    const localNotification = new firebase.notifications.Notification(settings)
+      .setNotificationId(`${notification.activityId}-${Math.random()}`) // Any random ID
+      .setTitle(notification.activityName) // Title of the notification
+      .setData({
+        event_id: notification.eventId,
+        applet_id: notification.appletId,
+        activity_id: notification.activityId,
+        type: "event-alert"
+      })
+      .android.setPriority(firebase.notifications.Android.Priority.High) // set priority in Android
+      .android.setChannelId(AndroidChannelId) // should be the same when creating channel for Android
+      .android.setAutoCancel(true); // To remove notification when tapped on it
+
+    firebase.notifications()
+      .scheduleNotification(localNotification, {
+        fireDate: notification.date,
+        repeatInterval: 'day',
+        exact: true,
+      })
+      .catch(err => console.error(err));
+  })
+};
+
+const closeExistingNotifications = () => {
+  firebase.notifications().cancelAllNotifications();
+}
+
+// const buildNotification = async (activity) => {
+//   const title = Platform.OS === "android" ? "Daily Reminder" : "";
+//   const AndroidChannelId = 'MindLoggerChannelId';
+//   const notificationData = {
+//     event_id: 1,
+//     applet_id: activity.appletId.split('/').pop(),
+//     activity_id: activity.id.split('/').pop(),
+//     type: "event-alert"
+//   }
+//   const settings = { showInForeground: true };
+//   const notification = new firebase.notifications.Notification(settings)
+//     .setNotificationId(`${activity.id}-${Math.random()}`) // Any random ID
+//     .setTitle(title) // Title of the notification
+//     .setBody("This is a notification") // body of notification
+//     .setData(notificationData);
+
+
+//   notification.android.setPriority(firebase.notifications.Android.Priority.High) // set priority in Android
+//   notification.android.setChannelId(AndroidChannelId) // should be the same when creating channel for Android
+//   notification.android.setAutoCancel(true); // To remove notification when tapped on it
+
+//   return notification;
+// };
+
 export const downloadApplets = (onAppletsDownloaded = null) => (dispatch, getState) => {
   const state = getState();
   const auth = authSelector(state);
@@ -79,7 +167,7 @@ export const downloadApplets = (onAppletsDownloaded = null) => (dispatch, getSta
       if (loggedInSelector(getState())) {
         // Check that we are still logged in when fetch finishes
         const transformedApplets = applets
-          .filter(applet => !R.isEmpty(applet.items))
+          .filter((applet) => !R.isEmpty(applet.items))
           .map(transformApplet);
         dispatch(replaceApplets(transformedApplets));
         dispatch(downloadResponses(transformedApplets));
@@ -89,7 +177,7 @@ export const downloadApplets = (onAppletsDownloaded = null) => (dispatch, getSta
         }
       }
     })
-    .catch(err => console.warn(err.message))
+    .catch((err) => console.warn(err.message))
     .finally(() => {
       dispatch(setDownloadingApplets(false));
       // dispatch(scheduleAndSetNotifications());
@@ -97,7 +185,10 @@ export const downloadApplets = (onAppletsDownloaded = null) => (dispatch, getSta
     });
 };
 
-export const downloadTargetApplet = (appletId, cb = null) => (dispatch, getState) => {
+export const downloadTargetApplet = (appletId, cb = null) => (
+  dispatch,
+  getState
+) => {
   const state = getState();
   const auth = authSelector(state);
   dispatch(setDownloadingTargetApplet(true));
@@ -106,7 +197,7 @@ export const downloadTargetApplet = (appletId, cb = null) => (dispatch, getState
       if (loggedInSelector(getState())) {
         // Check that we are still logged in when fetch finishes
         const transformedApplets = [applet]
-          .filter(applet => !R.isEmpty(applet.items))
+          .filter((applet) => !R.isEmpty(applet.items))
           .map(transformApplet);
         if (transformedApplets && transformedApplets.length > 0) {
           const transformedApplet = transformedApplets[0];
@@ -121,10 +212,10 @@ export const downloadTargetApplet = (appletId, cb = null) => (dispatch, getState
         }
       }
     })
-    .catch(err => console.warn(err.message));
+    .catch((err) => console.warn(err.message));
 };
 
-export const acceptInvitation = inviteId => (dispatch, getState) => {
+export const acceptInvitation = (inviteId) => (dispatch, getState) => {
   const state = getState();
   const auth = authSelector(state);
   return acceptAppletInvite(auth.token, inviteId).then(() => {
@@ -133,7 +224,7 @@ export const acceptInvitation = inviteId => (dispatch, getState) => {
   });
 };
 
-export const declineInvitation = inviteId => (dispatch, getState) => {
+export const declineInvitation = (inviteId) => (dispatch, getState) => {
   const state = getState();
   const auth = authSelector(state);
 
@@ -147,7 +238,7 @@ export const declineInvitation = inviteId => (dispatch, getState) => {
     });
 };
 
-export const joinOpenApplet = appletURI => (dispatch, getState) => {
+export const joinOpenApplet = (appletURI) => (dispatch, getState) => {
   dispatch(setDownloadingApplets(true));
   const state = getState();
   const auth = authSelector(state);
@@ -160,13 +251,13 @@ export const joinOpenApplet = appletURI => (dispatch, getState) => {
     });
 };
 
-export const updateBadgeNumber = badgeNumber => (dispatch, getState) => {
+export const updateBadgeNumber = (badgeNumber) => (dispatch, getState) => {
   const state = getState();
   const token = state.user?.auth?.token;
   if (token) {
     postAppletBadge(token, badgeNumber)
       .then((response) => {
-        console.log('updateBadgeNumber success', response);
+        console.log("updateBadgeNumber success", response);
       })
       .catch((e) => {
         console.warn(e);
@@ -174,34 +265,34 @@ export const updateBadgeNumber = badgeNumber => (dispatch, getState) => {
   }
 };
 
-export const deactivateApplet = groupId => (dispatch, getState) => {
+export const deactivateApplet = (groupId) => (dispatch, getState) => {
   const state = getState();
   const auth = authSelector(state);
   removeApplet(auth.token, groupId).then(() => {
     dispatch(setCurrentApplet(null));
     dispatch(sync());
-    Actions.push('applet_list');
+    Actions.push("applet_list");
   });
 };
 
-export const removeAndDeleteApplet = groupId => (dispatch, getState) => {
+export const removeAndDeleteApplet = (groupId) => (dispatch, getState) => {
   const state = getState();
   const auth = authSelector(state);
   deleteApplet(auth.token, groupId).then(() => {
     dispatch(setCurrentApplet(null));
     dispatch(sync());
-    Actions.push('applet_list');
+    Actions.push("applet_list");
   });
 };
 
-export const getAppletResponseData = appletId => (dispatch, getState) => {
+export const getAppletResponseData = (appletId) => (dispatch, getState) => {
   const state = getState();
   const auth = authSelector(state);
   getLast7DaysData({
     authToken: auth.token,
     appletId,
   }).then((resp) => {
-    // console.log('response is', resp);
+    console.log("response is", resp);
     dispatch(saveAppletResponseData(appletId, resp));
   });
 };