+++ app/widgets/AudioStimulus/AudioPlayer.js
@@ -1,37 +1,37 @@
-import React from 'react';
-import { View, TouchableOpacity, StyleSheet } from 'react-native';
-import PropTypes from 'prop-types';
-import { Icon } from 'native-base';
-import { Player } from '@react-native-community/audio-toolkit';
-import { colors } from '../../themes/colors';
-import BaseText from '../../components/base_text/base_text';
+import React from "react";
+import { View, TouchableOpacity, StyleSheet } from "react-native";
+import PropTypes from "prop-types";
+import { Icon } from "native-base";
+import { Player } from "@react-native-community/audio-toolkit";
+import { colors } from "../../themes/colors";
+import BaseText from "../../components/base_text/base_text";
 
 const styles = StyleSheet.create({
   playButton: {
     borderRadius: 3,
     backgroundColor: colors.primary,
     padding: 16,
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
+    flexDirection: "row",
+    justifyContent: "center",
+    alignItems: "center",
     width: 160,
     marginRight: 16,
-    shadowColor: '#000',
+    shadowColor: "#000",
     shadowOffset: { width: 0, height: 2 },
     shadowOpacity: 0.2,
     shadowRadius: 5,
     elevation: 1,
   },
   buttonText: {
-    fontWeight: 'bold',
-    color: 'white',
+    fontWeight: "bold",
+    color: "white",
     fontSize: 18,
     marginLeft: 5,
     marginRight: 5,
   },
   container: {
-    flexDirection: 'row',
-    alignItems: 'center',
+    flexDirection: "row",
+    alignItems: "center",
   },
   infoText: {
     color: colors.tertiary,
@@ -48,7 +48,6 @@ export class AudioPlayer extends React.Component {
       playbackCount: 0,
     };
   }
-
   play = () => {
     const { source, onEnd } = this.props;
     const { playbackCount } = this.state;
@@ -61,12 +60,11 @@ export class AudioPlayer extends React.Component {
     });
 
     // Listen for end of track
-    p.on('ended', () => {
+    p.on("ended", () => {
       this.setState({ playing: false });
       onEnd();
     });
   };
-
   stop = () => {
     const { player } = this.state;
     this.setState({ playing: false });
@@ -90,7 +88,6 @@ export class AudioPlayer extends React.Component {
   render() {
     const { allowReplay } = this.props;
     const { playing, playbackCount } = this.state;
-
     if (playing && allowReplay) {
       return (
         <TouchableOpacity onPress={this.stop}>
@@ -107,8 +104,15 @@ export class AudioPlayer extends React.Component {
       return (
         <TouchableOpacity disabled>
           <View style={[styles.playButton, { opacity: 0.5 }]}>
-            <BaseText style={styles.buttonText} textKey="audio_player:playing" />
-            <Icon style={styles.buttonText} type="FontAwesome" name="volume-up" />
+            <BaseText
+              style={styles.buttonText}
+              textKey="audio_player:playing"
+            />
+            <Icon
+              style={styles.buttonText}
+              type="FontAwesome"
+              name="volume-up"
+            />
           </View>
         </TouchableOpacity>
       );