+++ app/widgets/MultiSelect.js
@@ -1,11 +1,11 @@
-import React, { Component } from 'react';
-import { View, Image } from 'react-native';
-import PropTypes from 'prop-types';
-import * as R from 'ramda';
-import { ListItem, Text } from 'native-base';
-import { CheckBox } from 'react-native-elements';
-import { getURL } from '../services/helper';
-import { colors } from '../themes/colors';
+import React, { Component } from "react";
+import { View, Image } from "react-native";
+import PropTypes from "prop-types";
+import * as R from "ramda";
+import { ListItem, Text } from "native-base";
+import { CheckBox } from "react-native-elements";
+import { getURL } from "../services/helper";
+import { colors } from "../themes/colors";
 
 export class MultiSelect extends Component {
   static isValid(value = [], { minValue = 1, maxValue = Infinity }) {
@@ -34,18 +34,18 @@ export class MultiSelect extends Component {
       value = [],
     } = this.props;
     return (
-      <View style={{ alignItems: 'stretch' }}>
+      <View style={{ alignItems: "stretch" }}>
         {itemList.map((item, index) => (
           <ListItem
-            style={{ width: '90%' }}
+            style={{ width: "90%" }}
             onPress={() => this.onAnswer(token ? item.name.en : item.value)}
             key={index}
           >
-            <View style={{ width: '85%' }}>
-              <View style={{ width: '100%', flexDirection: 'row' }}>
+            <View style={{ width: "85%" }}>
+              <View style={{ width: "100%", flexDirection: "row" }}>
                 {item.image ? (
                   <Image
-                    style={{ width: '20%', height: 64, resizeMode: 'contain' }}
+                    style={{ width: "20%", height: 64, resizeMode: "contain" }}
                     source={{ uri: getURL(item.image) }}
                   />
                 ) : (
@@ -54,29 +54,47 @@ export class MultiSelect extends Component {
                 {item.image ? (
                   <View
                     style={{
-                      marginLeft: '8%',
-                      maxWidth: '72%',
-                      justifyContent: 'center',
+                      marginLeft: "8%",
+                      maxWidth: "72%",
+                      justifyContent: "center",
                     }}
                   >
-                    <Text>{item.name.en} {token ? (item.value < 0 ? '(' + item.value + ')' : '(+' + item.value + ')') : ""}</Text>
+                    <Text>
+                      {item.name.en}{" "}
+                      {token
+                        ? item.value < 0
+                          ? "(" + item.value + ")"
+                          : "(+" + item.value + ")"
+                        : ""}
+                    </Text>
                   </View>
                 ) : (
                   <View
                     style={{
-                      marginLeft: '8%',
-                      maxWidth: '92%',
-                      justifyContent: 'center',
+                      marginLeft: "8%",
+                      maxWidth: "92%",
+                      justifyContent: "center",
                     }}
                   >
-                      <Text>{item.name.en} {token ? (item.value < 0 ? '(' + item.value + ')' : '(+' + item.value + ')') : ""}</Text>
+                    <Text>
+                      {item.name.en}{" "}
+                      {token
+                        ? item.value < 0
+                          ? "(" + item.value + ")"
+                          : "(+" + item.value + ")"
+                        : ""}
+                    </Text>
                   </View>
                 )}
               </View>
             </View>
-            <View style={{ width: '15%' }}>
+            <View style={{ width: "15%" }}>
               <CheckBox
-                checked={value && Array.isArray(value) && value.includes(token ? item.name.en : item.value)}
+                checked={
+                  value &&
+                  Array.isArray(value) &&
+                  value.includes(token ? item.name.en : item.value)
+                }
                 onPress={() => this.onAnswer(token ? item.name.en : item.value)}
                 checkedIcon="check-square"
                 uncheckedIcon="square-o"